Update burger.js
This is an update of the code that I used for my project. This is my css for the project that I need to finsh before class monday. I made two literal objects and place them into the div boxes that I made. I just need to find a way to print the words onto the div boxes to get an A on this test.